+++ CHANGELOG.md
@@ -13,6 +13,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [2.0.10] - 2024-07-11
+
+### Added
+
+Added IFileDialogManager, a wrapper for dalamud's file dialog manager, made to avoid font crashes
+DalaMock will now save and load a global game path and plugin config path
+
+### Fixed
+
+Block plugin loading when paths are invalid and stop paths being edit if plugin is running
+
+### Changed
+
+### Removed
+
 ## [2.0.9] - 2024-07-11
 
 ### Added